backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
openai
google-generativeai
anthropic
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
groq
